Fixed Percent margin test on IOS
